+++ ui-spa/src/components/case-management/transfer-materials/egressFolderContainer.module.scss
@@ -1,0 +1,7 @@
+.iconButtonWrapper {
+  display: flex;
+  align-items: center;
+}
+.fileName {
+  margin-left: 0.25rem;
+}
