+++ ui-spa/src/components/case-management/transfer-materials/index.module.scss
@@ -2,10 +2,7 @@
   display: flex;
   width: 100%;
   gap: 10px;
-  .iconButtonWrapper {
-    display: flex;
-    align-items: center;
-  }
+
   .titleWrapper {
     background: #f3f2f1;
     padding: 0.25rem 0.5rem;
@@ -20,9 +17,6 @@
   .egressContainer {
     width: 60%;
     border: 1px solid #b1b4b6;
-    & .fileName {
-      margin-left: 0.25rem;
-    }
 
     & :global(.govuk-checkboxes__input) {
       width: 30px;
@@ -42,7 +36,6 @@
     & :global(.govuk-table) {
       table-layout: fixed;
     }
-
     & :global(.govuk-table__header:nth-of-type(1)) {
       width: 1.25rem;
       & :global(.govuk-checkboxes__item) {
@@ -71,8 +64,5 @@
   .netappContainer {
     width: 40%;
     border: 1px solid #b1b4b6;
-    & :global(.govuk-table) {
-      // table-layout: fixed;
-    }
   }
 }
