+++ ui-spa/src/components/case-management/transfer-materials/index.tsx
@@ -1,18 +1,9 @@
 import { useEffect, useState, useMemo } from "react";
 import { useApi } from "../../../common/hooks/useApi";
 import { LinkButton, InsetText } from "../../govuk";
-import Checkbox from "../../common/Checkbox";
-import FolderNavigationTable from "../../common/FolderNavigationTable";
-import {
-  sortByStringProperty,
-  sortByDateProperty,
-} from "../../../common/utils/sortUtils";
-import FolderIcon from "../../../components/svgs/folder.svg?react";
-import FileIcon from "../../../components/svgs/file.svg?react";
 import NetAppFolderContainer from "./NetAppFolderContainer";
-import { formatDate } from "../../../common/utils/formatDate";
-import { formatFileSize } from "../../../common/utils/formatFileSize";
 import { getEgressFolders, getNetAppFolders } from "../../../apis/gateway-api";
+import EgressFolderContainer from "./EgressFolderContainer";
 import styles from "./index.module.scss";
 
 type TransferMaterialsPageProps = {
@@ -24,10 +15,6 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
   egressWorkspaceId,
   netAppPath,
 }) => {
-  const [sortValues, setSortValues] = useState<{
-    name: string;
-    type: "ascending" | "descending";
-  }>();
   const [egressPathFolders, setEgressPathFolders] = useState<
     {
       folderName: string;
@@ -45,8 +32,7 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
   useEffect(() => {
     if (netAppPath) setNetAppFolderPath(netAppPath);
   }, [netAppPath]);
-  console.log("netAppFolderPath>>", netAppFolderPath);
-  console.log("netAppPath>>", netAppPath);
+
   const currentEgressFolder = useMemo(() => {
     return egressPathFolders[egressPathFolders.length - 1];
   }, [egressPathFolders]);
@@ -67,27 +53,6 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
     data: netAppData,
   } = useApi(getNetAppFolders, [netAppFolderPath], false);
 
-  const egressFolderData = useMemo(() => {
-    if (!egressData) return [];
-    if (sortValues?.name === "folder-name")
-      return sortByStringProperty(egressData, "name", sortValues.type);
-
-    if (sortValues?.name === "date-updated")
-      return sortByDateProperty(egressData, "dateUpdated", sortValues.type);
-
-    if (sortValues?.name === "file-size")
-      return sortByStringProperty(egressData, "filesize", sortValues.type);
-
-    return egressData;
-  }, [egressData, sortValues]);
-
-  const handleTableSort = (
-    sortName: string,
-    sortType: "ascending" | "descending",
-  ) => {
-    setSortValues({ name: sortName, type: sortType });
-  };
-
   const handleFolderPathClick = (path: string) => {
     const index = egressPathFolders.findIndex(
       (item) => item.folderPath === path,
@@ -100,7 +65,7 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
   };
 
   const handleFolderClick = (id: string) => {
-    const folderData = egressFolderData.find((item) => item.id === id);
+    const folderData = egressData!.find((item) => item.id === id);
     if (folderData)
       setEgressPathFolders((prevItems) => [
         ...prevItems,
@@ -117,10 +82,7 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
 
     if (id === "all-folders") {
       if (checked) {
-        updatedFolders = [
-          "all-folders",
-          ...egressFolderData.map((data) => data.id),
-        ];
+        updatedFolders = ["all-folders", ...egressData!.map((data) => data.id)];
       } else {
         updatedFolders = [];
       }
@@ -139,94 +101,6 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
 
   const handleGetFolderContent = (path: string) => {
     setNetAppFolderPath(path);
-  };
-
-  const getTableHeadData = () => {
-    return [
-      {
-        children: (
-          <>
-            <Checkbox
-              id={"all-folders"}
-              checked={isEgressFolderChecked("all-folders")}
-              onChange={handleCheckboxChange}
-              ariaLabel="Select all folders"
-            />
-          </>
-        ),
-        sortable: false,
-      },
-      {
-        children: <>Folder/file name</>,
-        sortable: true,
-        sortName: "folder-name",
-      },
-
-      {
-        children: <>Last modified date</>,
-        sortable: true,
-        sortName: "date-updated",
-      },
-      {
-        children: <>Size</>,
-        sortable: true,
-        sortName: "file-size",
-      },
-    ];
-  };
-
-  const getTableRowData = () => {
-    return egressFolderData.map((data) => {
-      return {
-        cells: [
-          {
-            children: (
-              <>
-                <Checkbox
-                  id={data.id}
-                  checked={isEgressFolderChecked(data.id)}
-                  onChange={handleCheckboxChange}
-                  ariaLabel="select folder"
-                />
-              </>
-            ),
-          },
-          {
-            children: (
-              <div className={styles.iconButtonWrapper}>
-                {data.isFolder ? (
-                  <>
-                    <FolderIcon />
-                    <LinkButton
-                      type="button"
-                      onClick={() => {
-                        handleFolderClick(data.id);
-                      }}
-                    >
-                      {data.name}
-                    </LinkButton>
-                  </>
-                ) : (
-                  <>
-                    <FileIcon />
-                    <span className={styles.fileName}>{data.name}</span>
-                  </>
-                )}
-              </div>
-            ),
-          },
-
-          {
-            children: <span>{formatDate(data.dateUpdated)}</span>,
-          },
-          {
-            children: (
-              <span>{data.filesize ? formatFileSize(data.filesize) : ""}</span>
-            ),
-          },
-        ],
-      };
-    });
   };
 
   useEffect(() => {
@@ -252,17 +126,17 @@ const TransferMaterialsPage: React.FC<TransferMaterialsPageProps> = ({
           <h3>Egress Inbound documents</h3>
         </div>
         <div className={styles.tableContainer}>
-          <FolderNavigationTable
-            tableName="egress"
-            folders={egressPathFolders}
-            loaderText="Loading folders from Egress"
-            folderResultsStatus={egressStatus}
-            folderResultsLength={egressFolderData.length}
-            handleFolderPathClick={handleFolderPathClick}
-            getTableRowData={getTableRowData}
-            getTableHeadData={getTableHeadData}
-            handleTableSort={handleTableSort}
-          />
+          {
+            <EgressFolderContainer
+              egressData={egressData}
+              egressDataStatus={egressStatus}
+              egressPathFolders={egressPathFolders}
+              handleFolderPathClick={handleFolderPathClick}
+              handleFolderClick={handleFolderClick}
+              handleCheckboxChange={handleCheckboxChange}
+              isEgressFolderChecked={isEgressFolderChecked}
+            />
+          }
         </div>
       </div>
     );
